requirements.txt
"discord"
python-dotenv
pydal
mangadexpy
schedule
asyncio